// src/models/IoK8sApimachineryPkgApisMetaV1APIVersions/types.ts
import { IoK8sApimachineryPkgApisMetaV1ServerAddressByClientCIDRModelOptions as IoK8sApimachineryPkgApisMetaV1ServerAddressByClientCIDR } from '../IoK8sApimachineryPkgApisMetaV1ServerAddressByClientCIDR';

export interface IoK8sApimachineryPkgApisMetaV1APIVersionsModelOptions {
  apiVersion?: string;
  kind?: string;
  serverAddressByClientCIDRs: Array<IoK8sApimachineryPkgApisMetaV1ServerAddressByClientCIDR>;
  versions: Array<string>;
}